refactor: Refactor Prarse Convert Unmarshal
