+++ tools/schema/main.go
@@ -19,6 +19,7 @@ import (
 
 	"github.com/iotaledger/wasp/tools/schema/generator"
 	"github.com/iotaledger/wasp/tools/schema/model"
+	wasp_yaml "github.com/iotaledger/wasp/tools/schema/model/yaml"
 	"gopkg.in/yaml.v3"
 )
 
@@ -236,10 +237,9 @@ func loadSchema(file *os.File) (s *model.Schema, err error) {
 			err = WriteYAMLSchema(schemaDef)
 		}
 	case ".yaml":
+		schemaDef = model.NewSchemaDef()
 		fileByteArray, _ := io.ReadAll(file)
-		var yamlSchemaDef model.YAMLSchemaDef
-		err = yaml.Unmarshal(fileByteArray, &yamlSchemaDef)
-		schemaDef = yamlSchemaDef.ToSchemaDef()
+		err = wasp_yaml.Unmarshal(fileByteArray, schemaDef)
 		if err == nil && *flagType == "convert" {
 			err = WriteJSONSchema(schemaDef)
 		}
